+++ src/librustc/session/mod.rs
@@ -5,7 +5,7 @@ use rustc_session::cgu_reuse_tracker::CguReuseTracker;
 use rustc_data_structures::fingerprint::Fingerprint;
 use rustc_data_structures::fx::{FxHashMap, FxHashSet};
 
-use crate::lint;
+use rustc_session::lint;
 use crate::session::config::{OutputType, PrintRequest, Sanitizer, SwitchWithOptPath};
 use crate::session::search_paths::{PathKind, SearchPath};
 use crate::util::common::{duration_to_secs_str, ErrorReported};
